@@ -55,6 +55,7 @@
     "@angular/compiler-cli": "^6.0.6",
     "@types/jasmine": "^2.8.8",
     "@types/jasmine-matchers": "^0.2.30",
+    "@types/jasminewd2": "^2.0.3",
     "@types/lodash": "^4.14.110",
     "@types/node": "^6.0.113",
     "@types/object-hash": "^1.2.0",
@@ -64,6 +65,7 @@
     "jasmine-core": "^2.99.1",
     "jasmine-marbles": "^0.3.1",
     "jasmine-spec-reporter": "~3.2.0",
+    "jasminewd2": "^2.2.0",
     "karma": "~1.4.1",
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
@@ -72,10 +74,10 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-jasmine-matchers": "^3.7.0",
     "node-sass": "^4.9.0",
-    "protractor": "~5.1.0",
+    "protractor": "^5.4.0",
     "ts-mockito": "^2.3.0",
-    "ts-node": "~2.0.0",
+    "ts-node": "^7.0.1",
     "tslint": "~5.7.0",
-    "typescript": "~2.7.2"
+    "typescript": "~2.9.2"
   }
 }
