@@ -1,14 +1,11 @@
 import { Injectable } from '@angular/core';
 import { ApiService, RouteParams } from '../api.service';
-import { combineLatest, merge, Observable, of, pipe } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { Operation } from 'fast-json-patch';
 import { PublicMembersInterface } from '../../../models/public-members.interface';
 import { ApiRoute } from 'api';
 import { DataStore } from './data-store';
-import { combineAll, flatMap, map, mergeAll, mergeMap, tap } from 'rxjs/operators';
-import { flatten } from '@angular/compiler';
-import { fromArray } from 'rxjs/internal/observable/fromArray';
-
+import { combineAll, flatMap, map, mergeMap, switchMap, tap } from 'rxjs/operators';
 
 
 export type ReFetchQuery = string | {params: RouteParams, query: RouteParams} | RouteParams;
@@ -28,28 +25,35 @@ export class DataService implements PublicMembersInterface<ApiService> {
       return this.dataStore.getItem(route.typeName, itemId).subject$.asObservable();
     }
     else if (itemId /* assuming single entity is requested */) {
-      return this.api.getData<T>(route, params, query).pipe( mergeMap(d => {
-        const id = this.dataStore.addItem(route.typeName, d);
-        return this.dataStore.getByStoreId(id).subject$.asObservable();
-      }));
+      const id = this.dataStore.addItem(route.typeName, {id: itemId}, true);
+      const storeItem = this.dataStore.getByStoreId(id);
+      storeItem.inFlight = this.api.getData<T>(route, params, query);
+      return storeItem.inFlight;
     }
     else /* assuming multiple entities are requested */ {
-      return <any> this.api.getData<T[]>(route, params, query).pipe(map(d => {
-        let storeIds: string[];
-        if (d instanceof Array) {
-          storeIds = d.map(data => {
-            return this.dataStore.addItem(route.typeName, data);
-          });
-        }
-        else {
-          storeIds = [this.dataStore.addItem(route.typeName, d)];
-        }
-        return storeIds;
-      }),
-        flatMap((ids) => {
-          return ids.map(id => this.dataStore.getByStoreId(id).subject$.asObservable());
-        }),
-        combineAll()
+      return this.api.getData<T[]>(route, params, query).pipe<T>(
+        switchMap((coll) => {
+          return coll.length !== 0 ?
+            of(coll).pipe(
+              map(d => {
+                let storeIds: string[];
+                if (d instanceof Array) {
+                  storeIds = d.map(data => {
+                    return this.dataStore.addItem(route.typeName, data);
+                  });
+                }
+                else {
+                  storeIds = [this.dataStore.addItem(route.typeName, d)];
+                }
+                return storeIds;
+              }),
+              flatMap((ids) => {
+                return ids.map(id => this.dataStore.getByStoreId(id).subject$.asObservable());
+              }),
+              combineAll()
+            ) :
+            of([]);
+        })
       );
     }
   }
@@ -59,12 +63,7 @@ export class DataService implements PublicMembersInterface<ApiService> {
     if (itemId && this.dataStore.checkIsInStore(route.typeName, itemId)) {
       const storeItem = this.dataStore.getItem(route.typeName, itemId);
       storeItem.inFlight = this.api.postData<T>(route, params, body, query);
-      return storeItem.inFlight.pipe(
-        mergeMap(data => {
-          storeItem.subject$.next(data);
-          return storeItem.subject$.asObservable();
-        })
-      );
+      return storeItem.inFlight
     }
     else if (itemId) {
       return this.api.postData<T>(route, params, body, query).pipe( mergeMap(d => {
@@ -73,17 +72,13 @@ export class DataService implements PublicMembersInterface<ApiService> {
       }));
     }
     else {
-      const storeId = this.dataStore.addItem(route.typeName, body);
+      const storeId = this.dataStore.addItem(route.typeName, body, true);
       const storeItem = this.dataStore.getByStoreId(storeId);
 
       storeItem.inFlight = this.api.postData<T>(route, params, body, query);
-      return storeItem.inFlight.pipe(
-        mergeMap(data => {
-          storeItem.permanentId = data.id;
-          storeItem.subject$.next(data);
-          return storeItem.subject$.asObservable();
-        })
-      );
+      return storeItem.inFlight.pipe(tap(data => {
+        storeItem.permanentId = data.id;
+      }));
     }
   }
 
@@ -91,31 +86,11 @@ export class DataService implements PublicMembersInterface<ApiService> {
     const itemId = getItemIdFromParams(params, route.typeName);
     const storeItem = this.dataStore.getItem(route.typeName, itemId);
     storeItem.inFlight = this.api.patchData<T>(route, params, operations, query);
-    return storeItem.inFlight.pipe(
-      mergeMap(data => {
-        storeItem.subject$.next(data);
-        return storeItem.subject$.asObservable();
-      })
-    );
+    return storeItem.inFlight;
   }
 
   deleteData<T>(route: ApiRoute, params: RouteParams, query?: RouteParams): Observable<T> {
-    const request = this.api.deleteData<T>(route, params, query);
-    // request.toPromise().then(r => {
-    //   const refetchRequests = concat(refetch.map(ref => this.reFetch(ref)));
-    // });
-    // const identifier = this.getIdentifier(route, params, query);
-    //
-    // const storeSubject = this.getFromStore<T>(identifier, {route, params, query}, true);
-    //
-    // if (storeSubject) {
-    //   request.subscribe(d => {
-    //     storeSubject.next(d);
-    //     storeSubject.complete();
-    //   }, e => storeSubject.error(e));
-    // }
-
-    return request;
+      return this.api.deleteData<T>(route, params, query);
   }
 
 }
