@@ -1,4 +1,4 @@
-import { ErrorHandler, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ApiService } from './api.service';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
@@ -12,7 +12,6 @@ import { DataService } from './data-service/data.service';
 import { NgxJsonViewerModule } from 'ngx-json-viewer';
 import { HotKeysService } from './hot-keys.service';
 import { DisabledContentComponent } from './disabled-content/disabled-content.component';
-import { DynamicStateErrorHandler } from './dynamic-state/dynamic-state-error.handler';
 
 const providers: Provider[] = [
   ApiService,
@@ -20,10 +19,6 @@ const providers: Provider[] = [
     provide: HTTP_INTERCEPTORS,
     useClass: CredentialsInterceptor,
     multi: true
-  },
-  {
-    provide: ErrorHandler,
-    useClass: DynamicStateErrorHandler
   },
   CookieService,
   DataService,
