@@ -1,115 +1,197 @@
-import { Component, EventEmitter, forwardRef, Input, OnInit, Output } from '@angular/core';
-import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
+import { Component, ContentChild, forwardRef, Input, OnDestroy, OnInit, TemplateRef } from '@angular/core';
+import {
+  AbstractControl,
+  ControlValueAccessor,
+  NG_VALIDATORS,
+  NG_VALUE_ACCESSOR,
+  ValidationErrors,
+  Validator,
+} from '@angular/forms';
 import { noop } from 'rxjs';
+import { Selectable, SelectionState } from '../../../models/selectable';
 import * as _ from 'lodash';
-import { ResponseBody } from '../../../api-models/responseBody';
 
 
-export const MULTIPLE_INPUT_CONTROL_VALUE_ACCESSOR: any = {
+let instanceCount = 0;
+
+export const MULTIPLE_INPUT_CONTROL_VALUE_ACCESSOR = {
   provide: NG_VALUE_ACCESSOR,
   useExisting: forwardRef(() => MultipleInputComponent),
-  multi: true,
+  multi: true
 };
 
-export enum MultipleInputTypes {
-  checkbox = 'checkbox',
-  radio = 'radio',
-  select = 'select'
-}
-
-interface MultipleInputResponseBody {
-  value: ResponseBody;
-  selected: ResponseBody;
-  isOther: boolean;
-}
+export const MULTIPLE_INPUT_VALIDATORS = {
+  provide: NG_VALIDATORS,
+  useExisting: forwardRef(() => MultipleInputComponent),
+  multi: true
+};
 
 @Component({
   selector: 'app-multiple-input',
   templateUrl: './multiple-input.component.html',
   styleUrls: ['./multiple-input.component.scss'],
-  providers: [MULTIPLE_INPUT_CONTROL_VALUE_ACCESSOR],
+  providers: [MULTIPLE_INPUT_CONTROL_VALUE_ACCESSOR, MULTIPLE_INPUT_VALIDATORS]
 })
-export class MultipleInputComponent implements OnInit, ControlValueAccessor {
-  MultipleInputTypes = MultipleInputTypes;
-  @Input() min: number;
-  @Input() max: number;
-  @Input() name: string;
-  @Input() type: MultipleInputTypes;
+export class MultipleInputComponent implements OnInit, OnDestroy, ControlValueAccessor, Validator {
 
+  instanceNumber: number;
 
-  // Placeholders for the callbacks which are later provided
-  // by the Control Value Accessor
-  private onTouchedCallback: () => void = noop;
-  private onChangeCallback: (_: any) => void = noop;
+  @ContentChild(TemplateRef)
+  public itemTemplate: TemplateRef<any>;
 
-
-  private _innerValue: ResponseBody[];
-  private _options: MultipleInputResponseBody[];
-
+  private _min: number;
   @Input()
-  get value(): ResponseBody[]{
-    return this._innerValue || [];
+  set min(min: number) {
+    this._min = min;
+    this.onValidatorChangeCallback();
   }
-  set value(v: ResponseBody[]) {
-    if (v) {
-      this._innerValue = [...v];
-      this.onChangeCallback(v);
+  get min(): number {
+    return this._min;
+  }
+
+  private _max: number;
+  @Input()
+  set max (max: number){
+    this._max = max;
+    this.onValidatorChangeCallback();
+  }
+  get max(): number {
+    return this._max;
+  }
+
+  private selectableOptions: Selectable<any>[] = [];
+  @Input()
+  set options(opts: any[]){
+    if (opts) {
+      opts.forEach((opt, index) => {
+        const identicalOptionIndex = opts.findIndex((o, i) => i !== index && _.isEqual(o, opt));
+        if (identicalOptionIndex >= 0) {
+         throw new Error(`Identical options found at [${index}] and [${identicalOptionIndex}]`);
+        }
+      });
+
+      this.selectableOptions = opts.map(o => ({
+        item: o,
+        state: this.getOptionState(o)
+      }));
+
+      this.onChangeCallback(this.value);
+    }
+    else {
+      this.selectableOptions = [];
+      this.value = [];
     }
   }
-
+  get options(): any[] {
+    return this.selectableOptions.map(o => o.item)
+  }
 
   @Input()
-  set options(o: ResponseBody[]) {
-    if (o) {
-      this._options = o.map(opt => {
-        return {
-          value: opt,
-          selected: {content: this.isOptionSelected(opt), origin: opt.origin, isOriginal: opt.isOriginal },
-          isOther: false,
-        };
-      });
+  set value(items: any[]) {
+    this.writeValue(items);
+  }
+  get value(): any[] {
+    return this.selectableOptions.filter(o => o.state === SelectionState.ON).map(o => o.item);
+  }
+
+  @Input()
+  multiSelect = true;
+
+  selectionStates = SelectionState;
+
+  private focused: any;
+  private onTouchedCallback: Function = noop;
+  private onChangeCallback: Function = noop;
+  private onValidatorChangeCallback: Function = noop;
+  private disabled: boolean;
+
+  constructor(
+  ) {
+    instanceCount ++;
+    this.instanceNumber = instanceCount;
+  }
+
+  private getOptionState(opt: any): SelectionState {
+    const selectableOption = this.selectableOptions.find(s => _.isEqual(opt, s.item));
+    if (selectableOption) {
+      return selectableOption.state;
+    }
+    return SelectionState.OFF;
+  }
+
+  ngOnInit() {
+  }
+
+  ngOnDestroy(): void {
+  }
+
+  onCheckboxChanged(event: Event, at: number) {
+    const input = (<HTMLInputElement>event.target);
+    if (!this.max ||
+      !input.checked ||
+      this.value.length < this.max) {
+      this.selectableOptions[at].state = input.checked ?
+        SelectionState.ON :
+        input.indeterminate ?
+          SelectionState.IND : SelectionState.OFF;
+
+      this.writeValue(this.value);
+    }
+    else {
       this.onTouchedCallback();
     }
   }
 
-
-  isOptionSelected(option: ResponseBody): boolean {
-    if (this.value) {
-      return !!this.value.find(r => _.isEqual(r, option));
-    }
+  onRadioChanged(event: Event, at: number) {
+    this.writeValue([this.selectableOptions[at].item]);
   }
 
-  writeValue(v: ResponseBody[]): void {
-    this.value = v;
-  }
-
-  onSelectionChanged(e: Event, option: MultipleInputResponseBody) {
-    const currentValue = [...this.value];
-    const state = (<HTMLInputElement>e.target).checked;
-    if (state && !this.isOptionSelected(option.value)) {
-      currentValue.push(option.value)
-    }
-    else if (!state && this.isOptionSelected(option.value)) {
-      const selectedIndex = currentValue.findIndex(o => _.isEqual(o, option.value))
-      currentValue.splice(selectedIndex, 1);
-    }
-    this.writeValue(currentValue);
-  }
-
-  // From ControlValueAccessor interface
-  registerOnChange(fn: any) {
+  registerOnChange(fn: any): void {
     this.onChangeCallback = fn;
   }
-
-  // From ControlValueAccessor interface
-  registerOnTouched(fn: any) {
+  registerOnTouched(fn: any): void {
     this.onTouchedCallback = fn;
   }
 
-  constructor() { }
-
-  ngOnInit() {
-
+  setDisabledState(isDisabled: boolean): void {
+    this.disabled = isDisabled;
   }
 
+  onBlur(value: any[]) {
+    this.focused = null;
+    this.onTouchedCallback(value);
+  }
+
+  onFocus(value: any) {
+    this.focused = value;
+    this.onTouchedCallback(value);
+  }
+
+  writeValue(selected: any[]): void {
+    selected = selected || [];
+    const options = this.options;
+    const selectedIndexes = selected.map(opt => options.findIndex(o => _.isEqual(opt, o)));
+    this.selectableOptions.forEach((opt, index) => {
+      opt.state = selectedIndexes.includes(index) ? SelectionState.ON : SelectionState.OFF;
+    });
+    this.onChangeCallback(selected);
+  }
+
+  registerOnValidatorChange(fn: () => void): void {
+    this.onValidatorChangeCallback = fn;
+  }
+
+  validate(c: AbstractControl): ValidationErrors | null {
+    const errors: ValidationErrors = {};
+
+    if (this.min && this.value.length < this.min) {
+      errors.min = true;
+    }
+
+    if (this.max && this.value.length > this.max) {
+      errors.max = true;
+    }
+
+    return Object.keys(errors).length ? errors : null;
+  }
 }
