@@ -1,7 +1,7 @@
 export interface Sortable<T> {
   order: number;
   item: T;
-  onChange: (value: any[]) => void;
+  onOrderChange: (value: any[]) => void;
   onBlur: (value: any) => void;
   onFocus: (value: any) => void;
 }
